requirements.txt
fastapi==0.100.0
uvicorn==0.23.0
python-dotenv==1.0.1
supabase==2.4.3
pydantic==1.10.13
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.24.1
cloudinary==1.36.0
twilio==8.10.0
email-validator==2.0.0
